Moved StripTypeInfo from StringExtensions to TypeInfoHelper
