+++ tests/trx2junit.Core.Tests/Internal/TypeInfoHelperTests/StripTypeInfo.cs
@@ -3,7 +3,7 @@
 using System.Collections.Generic;
 using NUnit.Framework;
 
-namespace gfoidl.Trx2Junit.Core.Tests.Extensions.StringExtensionsTests;
+namespace gfoidl.Trx2Junit.Core.Tests.Internal.TypeInfoHelperTests;
 
 [TestFixture]
 public class StripTypeInfo
@@ -19,13 +19,13 @@ public class StripTypeInfo
         yield return new TestCaseData(null).Returns(null);
         yield return new TestCaseData("").Returns("");
 
-        yield return new TestCaseData("Method1")                .Returns("Method1");
+        yield return new TestCaseData("Method1").Returns("Method1");
         yield return new TestCaseData("Method1(arg: { a = 3 })").Returns("Method1(arg: { a = 3 })");
 
-        yield return new TestCaseData("Class1.Method1")                .Returns("Method1");
+        yield return new TestCaseData("Class1.Method1").Returns("Method1");
         yield return new TestCaseData("Class1.Method1(arg: { a = 3 })").Returns("Method1(arg: { a = 3 })");
 
-        yield return new TestCaseData("SimpleUnitTest.Class1.Method1")                .Returns("Method1");
+        yield return new TestCaseData("SimpleUnitTest.Class1.Method1").Returns("Method1");
         yield return new TestCaseData("SimpleUnitTest.Class1.Method1(arg: { a = 3 })").Returns("Method1(arg: { a = 3 })");
     }
 }
